housing: css + refactor
