--- frontend/src/lib/housing/components/templates/card/Card.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-	import { cn } from '$lib/housing/utils'
-
-	const { class: className = '', ...rest } = $props()
-</script>
-
-<div
-	class={cn('rounded-lg border bg-card text-card-foreground shadow-sm', className)}
-	{...rest}
-></div>
