+++ frontend/src/lib/housing/housing.css
@@ -44,6 +44,7 @@ body {
 		--sidebar-border: 220 13% 91%;
 		--sidebar-ring: 217.2 91.2% 59.8%;
 	}
+	
 	.dark {
 		--background: 0 0% 3.9%;
 		--foreground: 0 0% 98%;
@@ -78,6 +79,7 @@ body {
 		--sidebar-border: 240 3.7% 15.9%;
 		--sidebar-ring: 217.2 91.2% 59.8%;
 	}
+
 	* {
 		border-color: hsl(var(--border));
 	}
